+++ .storybook/mock-content/extraLargePromo.js
@@ -1,62 +1,63 @@
+// eslint-disable-next-line import/prefer-default-export
 export const extraLargePromo = {
-  "_id": "K2FFSIYSVRC7HIBZBXLQ2CUPTY",
-  "credits": {
-    "by": [
+  _id: 'K2FFSIYSVRC7HIBZBXLQ2CUPTY',
+  credits: {
+    by: [
       {
-        "_id": "saracarothers",
-        "additional_properties": {
-          "original": {
-            "byline": "Sara Lynn Carothers"
-          }
+        _id: 'saracarothers',
+        additional_properties: {
+          original: {
+            byline: 'Sara Lynn Carothers',
+          },
         },
-        "name": "Sara Carothers",
-        "type": "author",
-        "url": "/author/sara-carothers/"
-      }
-    ]
-  },
-  "description": {
-    "basic": "This story has only a basic promo image (a large rock on the beach). This image should be seen in promo blocks, as well as in the Lead Art spot on the article page, and as the social promo image."
-  },
-  "display_date": "2021-04-23T21:13:31.477Z",
-  "headlines": {
-    "basic": "basic promo image"
-  },
-  "label": {
-    "basic": {
-      "display": true,
-      "text": "Free"
-    }
-  },
-  "owner": {
-    "sponsored": false
-  },
-  "promo_items": {
-    "basic": {
-      "resized_params": {
-        "400x225": "Ajb8uwIAZ0XSQ8IPqWGLLW7NifQ=filters:format(jpg):quality(70)/",
-        "400x267": "_iFZPwsbU19A0A9D1xRlPLhVx8A=filters:format(jpg):quality(70)/",
-        "400x300": "i-eI-H-xOkI6EyCk-Ph98hBjjRg=filters:format(jpg):quality(70)/",
-        "600x338": "cwoIRHNCzgkrRQCiW3SogwZS1AY=filters:format(jpg):quality(70)/",
-        "600x400": "WxcB_Wff18LDqZt1wd3JH8vPTZI=filters:format(jpg):quality(70)/",
-        "600x450": "LRgGYobwJDQZyfeEhtalyrn3Ad0=filters:format(jpg):quality(70)/",
-        "800x450": "vvQrhtG0ck1TcO519Mo2dZQCRTE=filters:format(jpg):quality(70)/",
-        "800x533": "LQ2EhhmbD5brfMTBNwl_PT8Tkdg=filters:format(jpg):quality(70)/",
-        "800x600": "AjVgpLNL9WYSLNei-Ks_kwDEa7Q=filters:format(jpg):quality(70)/"
+        name: 'Sara Carothers',
+        type: 'author',
+        url: '/author/sara-carothers/',
       },
-      "type": "image",
-      "url": "https://cloudfront-us-east-1.images.arcpublishing.com/corecomponents/RYNL5SXYCVDH7LCJPXAHYF3DJI.JPG"
-    }
+    ],
   },
-  "type": "story",
-  "website_url": "/test/promo-image-variations/2021/04/23/basic-promo-image/",
-  "websites": {
-    "the-gazette": {
-      "website_section": {
-        "_id": "/test/promo-image-variations",
-        "name": "Promo image variations"
+  description: {
+    basic: 'This story has only a basic promo image (a large rock on the beach). This image should be seen in promo blocks, as well as in the Lead Art spot on the article page, and as the social promo image.',
+  },
+  display_date: '2021-04-23T21:13:31.477Z',
+  headlines: {
+    basic: 'basic promo image',
+  },
+  label: {
+    basic: {
+      display: true,
+      text: 'Free',
+    },
+  },
+  owner: {
+    sponsored: false,
+  },
+  promo_items: {
+    basic: {
+      resized_params: {
+        '400x225': 'Ajb8uwIAZ0XSQ8IPqWGLLW7NifQ=filters:format(jpg):quality(70)/',
+        '400x267': '_iFZPwsbU19A0A9D1xRlPLhVx8A=filters:format(jpg):quality(70)/',
+        '400x300': 'i-eI-H-xOkI6EyCk-Ph98hBjjRg=filters:format(jpg):quality(70)/',
+        '600x338': 'cwoIRHNCzgkrRQCiW3SogwZS1AY=filters:format(jpg):quality(70)/',
+        '600x400': 'WxcB_Wff18LDqZt1wd3JH8vPTZI=filters:format(jpg):quality(70)/',
+        '600x450': 'LRgGYobwJDQZyfeEhtalyrn3Ad0=filters:format(jpg):quality(70)/',
+        '800x450': 'vvQrhtG0ck1TcO519Mo2dZQCRTE=filters:format(jpg):quality(70)/',
+        '800x533': 'LQ2EhhmbD5brfMTBNwl_PT8Tkdg=filters:format(jpg):quality(70)/',
+        '800x600': 'AjVgpLNL9WYSLNei-Ks_kwDEa7Q=filters:format(jpg):quality(70)/',
       },
-      "website_url": "/test/promo-image-variations/2021/04/23/basic-promo-image/"
-    }
-  }
-}
+      type: 'image',
+      url: 'https://cloudfront-us-east-1.images.arcpublishing.com/corecomponents/RYNL5SXYCVDH7LCJPXAHYF3DJI.JPG',
+    },
+  },
+  type: 'story',
+  website_url: '/test/promo-image-variations/2021/04/23/basic-promo-image/',
+  websites: {
+    'the-gazette': {
+      website_section: {
+        _id: '/test/promo-image-variations',
+        name: 'Promo image variations',
+      },
+      website_url: '/test/promo-image-variations/2021/04/23/basic-promo-image/',
+    },
+  },
+};