+++ _templates/content-source/new/readme.ejs.t
@@ -1,0 +1,18 @@
+---
+# helper funcs https://www.hygen.io/docs/templates/#helpers-and-inflections
+to: blocks/<%= h.inflection.dasherize(block_name) %>-content-source-block/README.md
+---
+# @wpmedia/<%= h.inflection.dasherize(block_name) %>-content-source-block
+_Please provide a 1-2 sentence description of what the @wpmedia/<%= h.inflection.dasherize(block_name) %>-content-source-block is and what it does._
+
+## ANS Schema
+Outline any schema information requirements necessary to know for ths block
+
+## Configurable Params
+None
+
+## TTL
+- `120`
+
+## Additional Considerations
+_This is optional. Please add an additional context that would be important to know in order to use this block._