+++ _templates/feature/feature/storybook-import.ejs.t
@@ -1,0 +1,6 @@
+---
+inject: true 
+to: blocks/<%= h.inflection.dasherize(block_name) %>-block/index.story.jsx
+before: "export default {"
+---
+import <%= h.changeCase.pascal(feature_name) %> from './features/<%= h.inflection.dasherize(feature_name) %>/default';