+++ blocks/header-nav-chain-block/index.story.jsx
@@ -5,7 +5,8 @@ export default {
   title: 'Chains/Header Nav Block',
   parameters: {
     chromatic: {
-      viewports: [320, 1200],
+      viewports: [320, 1200, 1800],
+      delay: 300,
     },
   },
 };
@@ -109,5 +110,63 @@ export const squareLogo = () => (
     signInOrder={1}
     primaryLogoPath="https://place-hold.it/100x100"
     primaryLogoAlt="Shows dimensions of 100 by 100 for tall testing purposes"
+  />
+);
+
+export const centerLogo = () => (
+  <PresentationalNav
+    backgroundColor="#fddede"
+    mediumBreakpoint={768}
+    closeDrawer={() => {}}
+    customFields={CUSTOM_FIELDS_BASE}
+    displayLinks
+    horizontalLinksHierarchy="horizontal-links"
+    isAdmin={false}
+    isSectionDrawerOpen={false}
+    logoAlignment="center"
+    menuButtonClickAction={() => { }}
+    navColor="light"
+    navColorClass="light"
+    navHeight={100}
+    scrollAdjustedNavHeight={100}
+    scrolled={false}
+    sectionAriaLabel="Menu des sections"
+    sections={[]}
+    showDotSeparators={false}
+    signInOrder={1}
+    primaryLogoPath="https://place-hold.it/100x100"
+    primaryLogoAlt="Shows dimensions of 100 by 100 for tall testing purposes"
+  />
+);
+
+const CUSTOM_FIELDS_BASE_2 = { ...CUSTOM_FIELDS_BASE };
+CUSTOM_FIELDS_BASE_2.leftComponentDesktop2 = 'none';
+CUSTOM_FIELDS_BASE_2.horizontalLinksHierarchy = 'tenLinks';
+CUSTOM_FIELDS_BASE_2.rightComponentDesktop1 = 'queryly';
+CUSTOM_FIELDS_BASE_2.rightComponentDesktop2 = 'menu';
+
+export const leftLogoWithLinks = () => (
+  <PresentationalNav
+    backgroundColor="#fddede"
+    mediumBreakpoint={768}
+    closeDrawer={() => {}}
+    customFields={CUSTOM_FIELDS_BASE_2}
+    displayLinks
+    horizontalLinksHierarchy="tenLinks"
+    isAdmin={false}
+    isSectionDrawerOpen={false}
+    logoAlignment="left"
+    menuButtonClickAction={() => { }}
+    navColor="light"
+    navColorClass="light"
+    navHeight={100}
+    scrollAdjustedNavHeight={100}
+    scrolled={false}
+    sectionAriaLabel="Menu des sections"
+    sections={[]}
+    showDotSeparators
+    signInOrder={1}
+    primaryLogoPath="https://place-hold.it/86x36"
+    primaryLogoAlt="Shows dimensions of 86 by 36."
   />
 );