+++ blocks/right-rail-advanced-block/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.17.0](https://github.com/WPMedia/arc-themes-blocks/compare/@wpmedia/right-rail-advanced-block@5.16.0...@wpmedia/right-rail-advanced-block@5.17.0) (2022-01-26)
+
+
+### Bug Fixes
+
+* Add layout section class to ensure spacing between elements ([8dec4a6](https://github.com/WPMedia/arc-themes-blocks/commit/8dec4a688bec6f14afa1729ae692eb6d9bc4a3e8))
+
+
+
+## 5.11.1-hotfix.0 (2021-04-23)
+
+
+
+## 5.10.3-hotfix.0 (2021-04-12)
+
+
+### Performance Improvements
+
+* Remove publishing of output type and other test files ([c95c97c](https://github.com/WPMedia/arc-themes-blocks/commit/c95c97c18f458219664d9276c6f3caa7af963a8b)), closes [#799](https://github.com/WPMedia/arc-themes-blocks/issues/799)
+
+
+
+# 5.10.0 (2021-02-17)
+
+
+
+
+
 # [5.16.0](https://github.com/WPMedia/fusion-news-theme-blocks/compare/@wpmedia/right-rail-advanced-block@5.15.0...@wpmedia/right-rail-advanced-block@5.16.0) (2021-09-15)